requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
bcrypt==4.0.1
beautifulsoup4==4.13.5
blinker==1.9.0
breadability==0.1.20
Brotli==1.1.0
caer==2.0.8
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==46.0.1
cycler==0.12.1
dataclasses-json==0.6.7
deep-translator==1.11.4
dnspython==2.8.0
docopt==0.6.2
ecdsa==0.19.1
faiss-cpu==1.12.0
fastapi==0.117.1
ffmpy==0.6.1
filelock==3.19.1
Flask==2.3.3
fonttools==4.56.0
frozenlist==1.7.0
fsspec==2025.9.0
gradio==5.47.0
gradio_client==1.13.2
greenlet==3.2.4
groovy==0.1.2
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
langsmith==0.4.30
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.3
marshmallow==3.26.1
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
mypy==1.9.0
mypy-extensions==1.0.0
networkx==3.5
nltk==3.8.1
numpy==1.26.4
opencv-contrib-python==4.9.0.80
orjson==3.11.3
packaging==24.2
pandas==2.2.1
pdfminer.six==20250506
pdfplumber==0.11.7
Pillow==10.0.1
pluggy==1.6.0
propcache==0.3.2
pymongo==4.10.1
dnspython==2.6.1
py4j==0.10.9.9
pyasn1==0.6.1
pycountry==24.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pymongo==4.5.0
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pyspark==4.0.0
pytesseract==0.3.10
pytest==7.4.2
pytest-flask==1.2.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
ruff==0.13.1
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==5.1.1
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
sumy==0.11.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2024.1
urllib3==2.5.0
uvicorn==0.24.0
websockets==15.0.1
Werkzeug==2.3.7
yarl==1.20.1
zstandard==0.25.0
